log/http: change behaviour of Request/RequestHeader
Before a request gets shallow copied in the private function
addToHTTPRequest but that feels somehow wrong. The burden of cloning
a request object must be shifted outside this package and document
here that Request and RequestHeader might trigger race conditions in
some situtation when a request has not been cloned.

Adding a function ShallowCloneRequest to allow easier cloning.